+++ irctc-dao/src/main/java/com/irctc/genericdao/CrudRepoGenericDao.java
@@ -1,13 +1,12 @@
 package com.irctc.genericdao;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CrudRepoGenericDao<T extends Serializable> extends CrudRepository<T, BigInteger> {
+public interface CrudRepoGenericDao<T extends Serializable> extends CrudRepository<T, Integer> {
